Fix translations conflicting between nested fields and translatable root fields of the same name (#51)
If a root field that is marked as translatable has the same name as a field in a nested form, the nested field was previously also appearing as translatable. On save, the value of the nested field would overwrite the value of the root field. This should fix this from occurring